backend/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# SSE
sse-starlette>=1.6.5

# Environment
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Vanna AI
vanna>=0.3.4

# LLM Integration
openai>=1.3.0
httpx>=0.25.0

# Database Connectivity
pyodbc>=4.0.39

# Vector Storage
chromadb>=0.4.15

# Data Processing
pandas>=2.0.0
plotly>=5.17.0

# File Processing
openpyxl>=3.1.0

python-multipart

matplotlib 